Add Thread-Safe Proxy Factory Workaround
Even when a list requires no sorting, List.sort keeps track of each sort
invocation as a change. When two different threads have a different
change modification count, then a ConcurrentModificationException is thrown

To work around that, this commit adds a thread-safe wrapper around
AuthorizationAdvisorProxyFactory and provides that wrapper to the
method advice.